+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog keeps track of all changes to the Packs SDK. We follow convention
 
 ## [Unreleased]
 
+### Added
+
+- Add initial `skill` definition to packs sdk
+
 ## [1.9.9] - 2025-07-03
 
 ### Changed
@@ -928,5 +932,4 @@ await myHelper(context);
 [1.9.5]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.9.5
 [1.9.8]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.9.8
 [1.9.7]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.9.7
-
 [1.9.9]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.9.9